open source publish 9eec484..b582ee1
X-Publish-Start-Commit: 9eec484c3701bd9aca8127d867820ed291be185a
X-Publish-End-Commit: b582ee1dd82af3429fd2f71e71bb650b2e5a68ae

---

Add nov2sbp, fix windows builds

+ Adds nov2sbp tool from Starling

+ Fixes windows builds by moving build off Rust nightly and having Rust handle stdin/stdout I/O on Windows

+ Some issues with code coverage popped up, seemingly unrelated to any specific code changes, it was decided that the code coverage targets that targeted the main functions (which operate on stdin/stdout) where functioning as non-deterministic fuzz tests (accidentally)-- so they were removed.

+ Introduces a fork of PROJ as a dependency (with sqlite embedded within: https://github.com/swift-nav/PROJ) to handle some geoid conversions, but only for nov2sbp-- this can be removed once we have a higher resolution grid for this in libswiftnav: https://swift-nav.atlassian.net/browse/INFRA-325

---

Disable build of nov2sbp by default

---

Remove -fPIC for libswiftnav

* Remove -fPIC for libswiftnav

* update PROJ

* update PROJ

---

Fix declaration of ubx2sbp_main

---

Update decoding of BDS SSR orbit messages to latest RTCM draft

---

Use HNR PVT to produce high rate heading

Replaces ORIENT_EULER output from NAV ATT parsing
while still using all message values from NAV ATT
except for heading, heading_acc and tow.

---

Remove duplicate declaration of readfn_ptr

---

static link c++ on windows

---

Mitigate WN issue for MSG_GNSS_TIME_OFFSET

---

Fix documentation and init n_sats field

---

workaround for mingw gcc bug

---
Remove PROJ submodule

---

Use full resolution geoid grid from libswiftnav instead of PROJ to compute geoidal separation values in nov2sbp

* Replace Transformer class in swiftnav_conversion_helpers.cc with call to get_geoid_offset() from libswiftnav
* Define GEOID_MODEL_15_MINUTE_RESOLUTION when compiling libswiftnav
* Remove proj dependency

---

Common compile flags

---

Replace Travis with GitHub Actions

+ Transition to GitHub Actions from Travis, mostly the same build structure as Travis, just adapted for GitHub Actions

+ This also attempts to try to refactor workflows into different files, but this introduces some coordination complications with GitHub Actions artifacts, so I'm not sure this was worth the effort... but maybe a useful technique for other types of automation.

Example release of artifacts: https://github.com/swift-nav/gnss-converters-private/releases/tag/v2021.03.02-test

---

Handle unknown week numbers form the M8L

+++ c/src/gnss_converters_extra/sbp_conv.c
@@ -30,7 +30,7 @@ struct sbp_conv_s {
 static s32 sbp_conv_cb(uint8_t *buf, uint16_t len, void *context) {
   fifo_t *fifo = context;
   assert(fifo != NULL);
-  return fifo_write(fifo, buf, len);
+  return (s32)fifo_write(fifo, buf, len);
 }
 
 sbp_conv_t sbp_conv_new() {
@@ -58,62 +58,63 @@ size_t sbp_conv(sbp_conv_t conv,
                 size_t wlen) {
   switch (type) {
     case SBP_MSG_BASE_POS_ECEF: {
-      sbp2rtcm_base_pos_ecef_cb(sender, rlen, rbuf, &conv->state);
+      sbp2rtcm_base_pos_ecef_cb(sender, (u8)rlen, rbuf, &conv->state);
       break;
     }
     case SBP_MSG_OBS: {
-      sbp2rtcm_sbp_obs_cb(sender, rlen, rbuf, &conv->state);
+      sbp2rtcm_sbp_obs_cb(sender, (u8)rlen, rbuf, &conv->state);
       break;
     }
     case SBP_MSG_OSR: {
-      sbp2rtcm_sbp_osr_cb(sender, rlen, rbuf, &conv->state);
+      sbp2rtcm_sbp_osr_cb(sender, (u8)rlen, rbuf, &conv->state);
       break;
     }
     case SBP_MSG_SSR_ORBIT_CLOCK: {
-      sbp2rtcm_sbp_ssr_orbit_clock_cb(sender, rlen, rbuf, &conv->state);
+      sbp2rtcm_sbp_ssr_orbit_clock_cb(sender, (u8)rlen, rbuf, &conv->state);
       break;
     }
     case SBP_MSG_SSR_PHASE_BIASES: {
-      sbp2rtcm_sbp_ssr_phase_biases_cb(sender, rlen, rbuf, &conv->state);
+      sbp2rtcm_sbp_ssr_phase_biases_cb(sender, (u8)rlen, rbuf, &conv->state);
       break;
     }
     case SBP_MSG_SSR_CODE_BIASES: {
-      sbp2rtcm_sbp_ssr_code_biases_cb(sender, rlen, rbuf, &conv->state);
+      sbp2rtcm_sbp_ssr_code_biases_cb(sender, (u8)rlen, rbuf, &conv->state);
       break;
     }
     case SBP_MSG_SSR_GRIDDED_CORRECTION_DEP_A: {
-      sbp2rtcm_sbp_ssr_gridded_correction_cb(sender, rlen, rbuf, &conv->state);
+      sbp2rtcm_sbp_ssr_gridded_correction_cb(
+          sender, (u8)rlen, rbuf, &conv->state);
       break;
     }
     case SBP_MSG_SSR_GRID_DEFINITION_DEP_A: {
-      sbp2rtcm_sbp_ssr_grid_definition_cb(sender, rlen, rbuf, &conv->state);
+      sbp2rtcm_sbp_ssr_grid_definition_cb(sender, (u8)rlen, rbuf, &conv->state);
       break;
     }
     case SBP_MSG_SSR_STEC_CORRECTION_DEP_A: {
-      sbp2rtcm_sbp_ssr_stec_correction_cb(sender, rlen, rbuf, &conv->state);
+      sbp2rtcm_sbp_ssr_stec_correction_cb(sender, (u8)rlen, rbuf, &conv->state);
       break;
     }
     case SBP_MSG_EPHEMERIS_GPS: {
-      sbp2rtcm_sbp_gps_eph_cb(sender, rlen, rbuf, &conv->state);
+      sbp2rtcm_sbp_gps_eph_cb(sender, (u8)rlen, rbuf, &conv->state);
       break;
     }
     case SBP_MSG_EPHEMERIS_GLO: {
-      sbp2rtcm_sbp_glo_eph_cb(sender, rlen, rbuf, &conv->state);
+      sbp2rtcm_sbp_glo_eph_cb(sender, (u8)rlen, rbuf, &conv->state);
       break;
     }
     case SBP_MSG_EPHEMERIS_BDS: {
-      sbp2rtcm_sbp_bds_eph_cb(sender, rlen, rbuf, &conv->state);
+      sbp2rtcm_sbp_bds_eph_cb(sender, (u8)rlen, rbuf, &conv->state);
       break;
     }
     case SBP_MSG_EPHEMERIS_GAL: {
-      sbp2rtcm_sbp_gal_eph_cb(sender, rlen, rbuf, &conv->state);
+      sbp2rtcm_sbp_gal_eph_cb(sender, (u8)rlen, rbuf, &conv->state);
       break;
     }
     case SBP_MSG_LOG: {
-      sbp2rtcm_sbp_log_cb(sender, rlen, rbuf, &conv->state);
+      sbp2rtcm_sbp_log_cb(sender, (u8)rlen, rbuf, &conv->state);
       break;
     }
     default: { break; }
   }
-  return fifo_read(&conv->fifo, wbuf, wlen);
+  return fifo_read(&conv->fifo, wbuf, (u8)wlen);
 }
